+++ src/proteus/workers/pt_zendnn.cpp
@@ -150,12 +150,16 @@ void PtZendnn::doAcquire(RequestParameters* parameters) {
 
   // Load the model
   std::string path;
-  if (parameters->has("model"))
+  if (parameters->has("model")) {
     path = parameters->get<std::string>("model");
-  else
+    if (!endsWith(path, ".pt")) {
+      path += ".pt";
+    }
+  } else {
     PROTEUS_LOG_ERROR(
       logger,
       "Model not provided");  // Ideally exit since model not provided
+  }
 
   auto module = torch::jit::load(
     path, torch::kCPU);  // Ideally exit if not able to read the model